Make operands and locals unsigned.

+++ frame.h
@@ -8,9 +8,9 @@
 
 typedef struct Frame {
   uint32_t n_locals;      /* computed at compile time             */
-  uint32_t *locals;       /* array of local variables             */
+  int32_t *locals;       /* array of local variables             */
   uint32_t n_operands;    /* computed at compile time             */
-  uint8_t *operands;      /* stack of operands                    */
+  int32_t *operands;      /* stack of operands                    */
   uint32_t i;             /* top of stack index                   */
   cp_info *cp;            /* reference to constant pool of class  */
 } Frame;
@@ -18,5 +18,5 @@ typedef struct Frame {
 void init_frame(Frame *f, uint32_t n_locals, uint32_t n_operands, cp_info *cp);
 void deinit_frame(Frame *f);
 
-void push_stack(Frame *f, uint32_t operand);
+void push_stack(Frame *f, int32_t operand);
 uint32_t pop_stack(Frame *f);
